+++ cmd/nvidia-ctk-installer/toolkit/toolkit_test.go
@@ -102,15 +102,6 @@ containerEdits:
           path: {{ .toolkitRoot }}/nvidia-cdi-hook
           args:
             - nvidia-cdi-hook
-            - create-soname-symlinks
-            - --folder
-            - /lib/x86_64-linux-gnu
-          env:
-            - NVIDIA_CTK_DEBUG=false
-        - hookName: createContainer
-          path: {{ .toolkitRoot }}/nvidia-cdi-hook
-          args:
-            - nvidia-cdi-hook
             - update-ldcache
             - --folder
             - /lib/x86_64-linux-gnu